Linux/requirements.txt
colorama==0.4.6
pycryptodome==3.20.0
